# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "EndpointConfigSummary": {
        "description": "端点配置概要——列表 / 创建 / 更新的响应体，密钥只报有无、绝不回内容。",
        "properties": {
          "api_format": {
            "description": "API 调用格式（一期仅 OpenAI Chat Completions）",
            "title": "Api Format",
            "type": "string"
          },
          "base_url": {
            "description": "端点地址",
            "title": "Base Url",
            "type": "string"
          },
          "has_api_key": {
            "description": "是否已存密钥（只报有无）",
            "title": "Has Api Key",
            "type": "boolean"
          },
          "id": {
            "description": "配置 ID（内部稳定身份，目录名即 ID，不随改名变化）",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "description": "是否为当前使用的配置",
            "title": "Is Active",
            "type": "boolean"
          },
          "model": {
            "description": "模型名",
            "title": "Model",
            "type": "string"
          },
          "name": {
            "description": "显示名（可改、允许重名）",
            "title": "Name",
            "type": "string"
          },
          "request_params": {
            "$ref": "#/components/schemas/EndpointRequestParams",
            "description": "已设置的请求参数（生成 + 传输）；未设置的键为 null"
          }
        },
        "required": [
          "id",
          "name",
          "base_url",
          "model",
          "api_format",
          "has_api_key",
          "is_active",
          "request_params"
        ],
        "title": "EndpointConfigSummary",
        "type": "object"
      },
      "EndpointCreateRequest": {
        "description": "POST /api/endpoints 的请求体——新增一套配置；api_key 缺省暂不配置（可用环境变量兜底）。",
        "properties": {
          "api_format": {
            "default": "openai-chat-completions",
            "title": "Api Format",
            "type": "string"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "description": "配置名（即目录名，1–64 字符、不含路径保留字符）",
            "title": "Name",
            "type": "string"
          },
          "request_params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EndpointRequestParams"
              },
              {
                "type": "null"
              }
            ],
            "description": "请求参数（生成 + 传输）；缺省 = 全不设（用默认值）"
          }
        },
        "required": [
          "name",
          "base_url",
          "model"
        ],
        "title": "EndpointCreateRequest",
        "type": "object"
      },
      "EndpointRequestParams": {
        "description": "一套端点配置的请求参数（生成 + 传输），随配置存于其 config.json。\n\n所有键都可缺省：null = 不设该参数（请求时用端点自身默认或工具内置默认）。",
        "properties": {
          "enable_thinking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "思考模式开关（一等参数）；null = 不传（跟随模型默认）。按 SiliconFlow / DashScope 官方口径进请求体顶层；仅部分模型支持（Qwen3.x / DeepSeek-V3.2+ / GLM / Kimi 等），不支持的模型会收到端点 400",
            "title": "Enable Thinking"
          },
          "extra_body": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "端点专有参数透传（openai SDK 的 extra_body，原样转发不解释）；null = 不传。厂商文档里的专有参数（如 reasoning_effort）放这里；与 enable_thinking 同名时一等参数优先",
            "title": "Extra Body"
          },
          "max_retries": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "失败自动重试次数；null = 用内置默认（2）",
            "title": "Max Retries"
          },
          "max_tokens": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "输出 token 上限；null = 不传",
            "title": "Max Tokens"
          },
          "temperature": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "采样温度；null = 不传（用端点默认）",
            "title": "Temperature"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "单次请求超时秒数；null = 用内置默认（120）",
            "title": "Timeout Seconds"
          },
          "top_p": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "核采样阈值；null = 不传",
            "title": "Top P"
          }
        },
        "title": "EndpointRequestParams",
        "type": "object"
      },
      "EndpointTestRequest": {
        "description": "POST /api/endpoints/test 的请求体：用表单当前值探测连通性（不必先保存）。",
        "properties": {
          "api_format": {
            "default": "openai-chat-completions",
            "description": "API 调用格式（一期仅 OpenAI Chat Completions；与受支持格式不符直接判失败）",
            "title": "Api Format",
            "type": "string"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "密钥；缺省先回落环境变量 DSF_API_KEY，再回落该配置已存密钥",
            "title": "Api Key"
          },
          "base_url": {
            "description": "端点根地址",
            "title": "Base Url",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "配置 ID（回落该配置已存密钥时用它定位；缺省 = 表单密钥与环境变量通道）",
            "title": "Id"
          },
          "model": {
            "description": "模型名",
            "title": "Model",
            "type": "string"
          },
          "request_params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EndpointRequestParams"
              },
              {
                "type": "null"
              }
            ],
            "description": "表单当前的高级参数（生成参数随探测一起发；传输参数由探测专用值覆盖）"
          }
        },
        "required": [
          "base_url",
          "model"
        ],
        "title": "EndpointTestRequest",
        "type": "object"
      },
      "EndpointTestResult": {
        "description": "POST /api/endpoints/test 的响应体：探测结果（HTTP 恒 200，成败看 ok）。",
        "properties": {
          "effective_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "本次探测实际发出的关键参数回显（model/stream/max_tokens/生成参数；探测专用传输参数不在此列）",
            "title": "Effective Params"
          },
          "latency_ms": {
            "description": "请求耗时（毫秒；未发出请求时为 0）",
            "title": "Latency Ms",
            "type": "number"
          },
          "message": {
            "description": "结果说明（失败时为分类后的可操作提示）",
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "description": "是否连通",
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "message",
          "latency_ms"
        ],
        "title": "EndpointTestResult",
        "type": "object"
      },
      "EndpointUpdateRequest": {
        "description": "PUT /api/endpoints/{name} 的请求体——api_key 缺省沿用已存密钥（不强迫重输）。",
        "properties": {
          "api_format": {
            "default": "openai-chat-completions",
            "title": "Api Format",
            "type": "string"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "new_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "新配置名（改名）；null = 不改名。字段更新成功后再执行改名",
            "title": "New Name"
          },
          "request_params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EndpointRequestParams"
              },
              {
                "type": "null"
              }
            ],
            "description": "请求参数（生成 + 传输）；缺省 = 沿用已有参数不变；提供 = 整体替换（未提供的参数键视为清除）"
          }
        },
        "required": [
          "base_url",
          "model"
        ],
        "title": "EndpointUpdateRequest",
        "type": "object"
      },
      "ErrorDetail": {
        "description": "错误响应的统一形状——所有域错误都长这样（`{\"detail\": \"一句话\"}`）。\n\n各端点在 `responses=` 里引用它，把错误路径写进 OpenAPI 契约：\n前端生成类型时才能覆盖错误体，而不是只看到成功路径。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "AI 打标工具（发图 + 指令产出 caption，支持迭代改写）",
    "title": "Dataset Factory",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/endpoints": {
      "get": {
        "description": "列出全部端点配置（按名称排序；密钥只报有无）。",
        "operationId": "list_all_api_endpoints_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EndpointConfigSummary"
                  },
                  "title": "Response List All Api Endpoints Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All",
        "tags": [
          "端点配置"
        ]
      },
      "post": {
        "description": "新增一套端点配置；当前没有生效配置时自动设为当前使用。",
        "operationId": "create_api_endpoints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointConfigSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "显示名不合法 / 字段为空 / API 格式暂未支持"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "端点配置"
        ]
      }
    },
    "/api/endpoints/test": {
      "post": {
        "description": "测试端点连通性：用表单当前值两档探测，不必先保存。\n\n密钥三通道（2026-09-21 审计定案，与实调的 resolve_api_key 对齐）：表单值 > 环境变量\nDSF_API_KEY > 该配置已存密钥——之前探测不认环境变量通道，会出现「实调能通、测试\n连接却报无密钥」的假故障。api_format 是真校验不是摆设：与受支持格式不符时直接给出\n可操作失败，不留「改了以为生效」的假字段。",
        "operationId": "test_connection_api_endpoints_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Connection",
        "tags": [
          "端点配置"
        ]
      }
    },
    "/api/endpoints/{cid}": {
      "delete": {
        "description": "删除一套端点配置（连同其密钥文件）。",
        "operationId": "remove_api_endpoints__cid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "配置不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "是当前使用中的配置（先切换到其他配置再删）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove",
        "tags": [
          "端点配置"
        ]
      },
      "put": {
        "description": "更新一套配置的端点字段；api_key 缺省沿用已存密钥、参数块缺省沿用已有参数。\n\n带 ``new_name`` 时改显示名（只写 config.json 的 name 字段；显示名允许重名，\n无冲突语义——身份是 ID）。",
        "operationId": "update_api_endpoints__cid__put",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndpointUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointConfigSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "字段为空 / API 格式暂未支持 / 显示名不合法"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "配置不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update",
        "tags": [
          "端点配置"
        ]
      }
    },
    "/api/endpoints/{cid}/activate": {
      "post": {
        "description": "把一套配置设为当前使用；对新请求立即生效。",
        "operationId": "activate_api_endpoints__cid__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "配置不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate",
        "tags": [
          "端点配置"
        ]
      }
    },
...
  }
}